# Optimize randBool() and randString()