I love this library, small but gets the job done! :)

I have tinkered a bit with the functions to generate bools and strings and made them a bit faster. Hopefully it will save someone some clocks when running tests.

The code speeds up bool generation 3-fold. I've ran some benchmarks on GCP VM's and my local MacBook:

> My MacBook Pro:
> 
> BenchmarkRandBool-8             75312272                14.5 ns/op
> BenchmarkRandBoolNew-8        235223112                5.06 ns/op
> BenchmarkRandStringStock-8       1439635               841 ns/op
> BenchmarkRandStringBuilder-8     1541736               775 ns/op
> 
> 8 core GCP E2:
> 
> BenchmarkRandBool-8             70003590                16.6 ns/op
> BenchmarkRandBoolNew-8        236062108                5.08 ns/op
> BenchmarkRandStringStock-8       1370116               874 ns/op
> BenchmarkRandStringBuilder-8     1484086               807 ns/op
> 
> 8 core GCP N2 (Intel Cascade Lake):
> 
> BenchmarkRandBool-8             91276062                13.1 ns/op
> BenchmarkRandBoolNew-8        301435986                3.98 ns/op
> BenchmarkRandStringStock-8       1743852               688 ns/op
> BenchmarkRandStringBuilder-8     1883102               638 ns/op
> 
> 8 core GCP N1 (Intel Skylake):
> 
> BenchmarkRandBool-8             71896552                16.5 ns/op
> BenchmarkRandBoolNew-8        239149131                5.03 ns/op
> BenchmarkRandStringStock-8       1375860               873 ns/op
> BenchmarkRandStringBuilder-8     1484378               808 ns/op

randString() will be 5-7% due to less data being moved around (thanks to strings.Builder)

randString() memory usage change:

> BenchmarkRandStringStock-8       1543179               772 ns/op              71 B/op          1 allocs/op
> BenchmarkRandStringBuilder-8     1639436               735 ns/op              48 B/op          1 allocs/op