@jake-ciolek 2020-06-04
Sure!

```
import (
	"math/rand"

	"testing"
	"strings"
)

func randBoolOld(r *rand.Rand) bool {
	if r.Int()&1 == 1 {
		return true
	}
	return false
}

func randBoolNew(r *rand.Rand) bool {
	return r.Int31()&(1<<30) == 0
}

func randStringOld(r *rand.Rand) string {
	n := r.Intn(20)
	runes := make([]rune, n)
	for i := range runes {
		runes[i] = unicodeRanges[r.Intn(len(unicodeRanges))].choose(r)
	}
	return string(runes)
}

func randStringBuilder(r *rand.Rand) string {
	n := r.Intn(20)
	sb := strings.Builder{}
        sb.Grow(n)
	for i := 0 ;i<n; i++ {
                sb.WriteRune(unicodeRanges[r.Intn(len(unicodeRanges))].choose(r))
	}
	
	return sb.String()
}

type int63nPicker interface {
	Int63n(int64) int64
}

type charRange struct {
	first, last rune
}

// choose returns a random unicode character from the given range, using the
// given randomness source.
func (cr charRange) choose(r int63nPicker) rune {
	count := int64(cr.last - cr.first + 1)
	return cr.first + rune(r.Int63n(count))
}

var unicodeRanges = []charRange{
	{' ', '~'},           // ASCII characters
	{'\u00a0', '\u02af'}, // Multi-byte encoded characters
	{'\u4e00', '\u9fff'}, // Common CJK (even longer encodings)
}

func BenchmarkRandBool(b *testing.B) {
	rs := rand.New(rand.NewSource(123))
	for i := 0; i < b.N; i++ {
		randBoolOld(rs)
    }
}

func BenchmarkRandBoolNew(b *testing.B) {
	rs := rand.New(rand.NewSource(123))
	for i := 0; i < b.N; i++ {
		randBoolNew(rs)
    }
}

func BenchmarkRandStringStock(b *testing.B) {
	rs := rand.New(rand.NewSource(123))
	for i := 0; i < b.N; i++ {
		randStringOld(rs)
    }
}

func BenchmarkRandStringBuilder(b *testing.B) {
	rs := rand.New(rand.NewSource(123))
	for i := 0; i < b.N; i++ {
		randStringBuilder(rs)
    }
}
```