@lavalamp 2020-06-04
Can you add the benchmarks to the _test.go file? (just one for Bool and one for String; no need for old/new--we can just track that over time)