@jake-ciolek 2020-06-04
Added benchmarks :)